add possibility to handle proxy info differently depending on upstream ip
Without such a possibility usage of this library is prone to security attacks
enabling simple faking of client ip addresses by accepting any proxy protocol
header regardless of whether the upstream source is allowed to send one or not.

This introduces the possibility to decide on each new incoming connection
whether the proxy header should be used, ignored, or reject the connection. It
also allows to reject the connection when a proxy header is required but not
send.