allow to specify IP ranges
This introduces the possibility to use IP ranges next to IP adresses for the whitelists. To ensure the provided values are valid they are parsed, invalid IP ranges or IP addresses are rejected. However, this requires LaxWhiteListPolicy() and StrictWhiteListPolicy() to return an error in such cases. To cope with that, similar to other Must* functions in the standard library, two new functions MustLaxWhiteListPolicy() and MustStrictWhiteListPolicy() which panic instead are introduced.