ci: use local Go toolchain and downgrade Go version
The GitHub Actions workflow defines:

```yaml
strategy:
  matrix:
    go: ['1.23', '1.24']
```

However, the `1.23` job still ends up running Go `1.24`. By default,
`actions/setup-go` leaves `GOTOOLCHAIN=auto`, causing go itself to fetch the
version specified in `go.mod` instead instead of sticking to exactly `1.23`.

This led to a false positive in CI: it appeared as if the matrix was testing
Go `1.23`, when in reality all jobs were using Go `1.24`.

To fix this, we need to force the use of the local Go toolchain in the CI test
workflow to ensure the use of the locally installed Go version and avoid
automatic downloads by `setup-go`.

Also downgrades the Go version in `go.mod` from `1.24` to `1.23` to ensure
compatibility with environments that do not yet support Go `1.24`.

Fixes #126